requirements.txt
langchain==0.0.247
Pillow==9.5.0
Pillow==10.0.0
Requests==2.31.0
streamlit==1.24.0
torch==2.0.1
transformers==4.31.0
